# docs/data/swagger.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
...
  "info": {
    "description": "API for DocsGPT",
    "title": "DocsGPT API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/approvals/{token}": {
      "get": {
        "description": "What an approval link asks: its title, question, details, options and whether it was decided. Public (the token is the credential); reading it changes nothing.",
        "operationId": "get_approval_link",
        "parameters": [
          {
            "description": "The link's token.",
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [],
        "tags": [
          "approvals"
        ],
        "x-pat-denied": true
      },
      "post": {
        "description": "Decide an approval link: {decision, comment?}, where decision is one of its options. The first decision wins and resumes the conversation that asked; later ones get 409. 404 for an unknown, expired or revoked link, 400 for an option it doesn't offer, 429 when rate limited.",
        "operationId": "post_approval_link",
        "parameters": [
          {
            "description": "The link's token.",
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [],
        "tags": [
          "approvals"
        ],
        "x-pat-denied": true
      }
    },
...
    "/api/background_jobs": {
      "get": {
        "description": "List the caller's background jobs of one conversation, newest first (no result bodies).",
        "operationId": "get_background_jobs",
        "parameters": [
          {
            "description": "The conversation whose jobs to list.",
            "in": "query",
            "name": "conversation_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "background_jobs"
        ],
        "x-pat-scopes": [
          "conversations:read"
        ]
      }
    },
    "/api/background_jobs/{job_id}": {
      "get": {
        "description": "One background job: its state, progress and, once finished, a result preview.",
        "operationId": "get_background_job",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "background_jobs"
        ],
        "x-pat-scopes": [
          "conversations:read"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "job_id",
          "required": true,
          "type": "string"
        }
      ]
    },
    "/api/background_jobs/{job_id}/cancel": {
      "parameters": [
        {
          "in": "path",
          "name": "job_id",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "description": "Ask a running background job to stop. It ends as cancelled once its work really stops; a step already under way may still complete.",
        "operationId": "post_cancel_background_job",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "background_jobs"
        ],
        "x-pat-scopes": [
          "conversations:write"
        ]
      }
    },
...
    "/api/monitors": {
      "get": {
        "description": "The caller's monitors, newest first: what each watches, its interval, last check, wakes left, expiry and links (never tokens or secrets).",
        "operationId": "get_monitors",
        "parameters": [
          {
            "description": "Only this conversation's monitors.",
            "in": "query",
            "name": "conversation_id",
            "type": "string"
          },
          {
            "description": "live (active and paused) or all (the default).",
            "in": "query",
            "name": "status",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "monitors"
        ],
        "x-pat-scopes": [
          "conversations:read"
        ]
      }
    },
    "/api/monitors/{monitor_id}": {
      "get": {
        "description": "One of the caller's monitors.",
        "operationId": "get_monitor_detail",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "monitors"
        ],
        "x-pat-scopes": [
          "conversations:read"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "monitor_id",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/api/monitors/{monitor_id}/{action}": {
      "parameters": [
        {
          "in": "path",
          "name": "monitor_id",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "description": "Pause, resume or cancel one of the caller's monitors. Cancelling revokes its links and any approval its source was given; resuming a polled monitor checks it again right away. 409 when the monitor is not in a state the action applies to.",
        "operationId": "post_monitor_action",
        "parameters": [
          {
            "description": "pause, resume or cancel",
            "in": "path",
            "name": "action",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "monitors"
        ],
        "x-pat-scopes": [
          "conversations:write"
        ]
      }
    },
...
    "/api/triggers/{token}": {
      "get": {
        "description": "A GET call, taken only by a link created with methods [\"POST\", \"GET\"]: the query parameters are the delivery (a dotted key nests). A HEAD, a prefetch (Purpose / Sec-Purpose: prefetch) or a known link-preview or crawler User-Agent gets 200 {ignored} and changes nothing. 202 when accepted, 405 for a POST-only link, 404 for an unknown, expired, revoked or used-up one, 429 when rate limited.",
        "operationId": "get_trigger_link",
        "parameters": [
          {
            "description": "The link's token.",
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [],
        "tags": [
          "triggers"
        ],
        "x-pat-denied": true
      },
      "post": {
        "description": "Deliver an event to a monitor's webhook trigger link. Public: the token is the credential. Accepts JSON, form or text up to TRIGGER_MAX_PAYLOAD_BYTES. Signed links verify Standard Webhooks (webhook-id/webhook-timestamp/webhook-signature, 5 minute tolerance), GitHub X-Hub-Signature-256, X-Signature: sha256=<hex>, Stripe-Signature (t=..,v1=.., 5 minute tolerance), Slack X-Slack-Signature with X-Slack-Request-Timestamp (5 minute tolerance; a url_verification challenge is answered with 200 {challenge}), a static token header (X-Webhook-Token unless the link names another) or Authorization: Bearer. Deduplicated on Idempotency-Key, webhook-id, X-GitHub-Delivery or the Stripe/Slack event id, else the body hash. 202 when accepted; 404 for an unknown, expired, revoked or used-up link; 401 for a bad signature; 413 for a large body; 429 past TRIGGER_RATE_PER_MINUTE.",
        "operationId": "post_trigger_link",
        "parameters": [
          {
            "description": "The link's token.",
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [],
        "tags": [
          "triggers"
        ],
        "x-pat-denied": true
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
...
  "swagger": "2.0",
...
}